Restructures folders structure for placeholder pages, for uniformity
