+++ frontend/src/features/documents/index.ts
@@ -1,0 +1,3 @@
+import Documents from "./Documents";
+
+export {Documents};
