+++ frontend/src/features/reports/index.ts
@@ -1,0 +1,3 @@
+import Reports from "./Reports";
+
+export {Reports};
